A4A Migrations commissions: reuse the shared consolidated stat card
Lift ConsolidatedStatCard out of the dashboard referrals folder into
client/dashboard/components/consolidated-stat-card so both features share
it, and make its info popover optional (referrals still passes one;
migrations does not). Repoint the referrals payout/consolidated views at
the shared location.

Rebuild MigrationsConsolidatedCommissions on the shared card: format the
commission amount with formatCurrency instead of a raw "$" string, and lay
the cards out with the same @wordpress/components Grid
(auto-fit, minmax(240px, 1fr)) the dashboard referrals uses instead of a
bespoke stylesheet. Keep the a8c-host content padding on the grid via the
consolidated-commissions class and drop the now-dead flex-direction row.
