+++ client/dashboard/agency/earn/referrals/payout-cards.tsx
@@ -2,8 +2,8 @@ import { formatCurrency } from '@automattic/number-formatters';
 import { __experimentalVStack as VStack } from '@wordpress/components';
 import { __, sprintf } from '@wordpress/i18n';
 import InlineSupportLink from '../../../components/inline-support-link';
+import ConsolidatedStatCard from '../../../components/consolidated-stat-card';
 import { formatDate } from '../../../utils/datetime';
-import ConsolidatedStatCard from './consolidated-stat-card';
 import useGetPayoutData from './hooks/use-payout-data';
 
 const AGENCY_EARNINGS_LEARN_MORE_LINK =
