+++ client/dashboard/components/consolidated-stat-card/style.scss
@@ -1,4 +1,4 @@
-.referrals-stat-card__popover-content {
+.consolidated-stat-card__popover-content {
 	width: min(320px, calc(100vw - 32px));
 	padding: 16px;
 	line-height: 1.4;
